Change the semantics of Image module
The Image module now makes no assumptions about how image variants are
stored.  Instead it expects the caller to supply 'imgset', which is hash
of DPR => image files.

The 'filename' and 'basename' attributes are replaced with 'id', which
might be anything, but the Gallery module will expect this to be image's
basename.

Also, 'caption' attribute must now be supplied by the caller, the Image
instance won't try to get it by itself.

This change is implement to allow the Image class handle gallery
thumbnails that use different semantics.