Add Overlay Scrollbars library to mitgate problems with scrollbars
Scrollbar appearing and disappearing messes up the mosaicing of the
gallery and I haven't figured a way to fix it.  Having the BODY overflow:
hidden makes the whole problem go away, but then scrolling the viewport
doesn't work anymore, hence the Overlay Scrollbars.  It's fairly heavy
library, so it would still be preferable to not need it.