Build an underpinning for deep linking feature
This reworks the loading of index.json so that we look for it at current
directory and also at parent directory (which would be the case for
deep-linked access).  Following Apache config sets up the
front-controller for this to work:

    <Directory /www/dev/gallery>
      RewriteEngine On
      RewriteBase "/dev/gallery/"
      RewriteRule "\.json$" - [L]
      RewriteCond %{REQUEST_FILENAME} !-f
      RewriteRule . /dev/gallery/gallery.html [L]
    </Directory>

Note, that JSON files must be excluded from the rewriting, since we have
to know when they are not present.  This makes it impossible to use the
FallbackResource statement for this.

Also note that this doesn't implement the actual deep linking!  Gallery
still only works when accessed from the gallery URL.