Implement history management
Every navigation within the gallery is pushed into browser history, so
Back/Forward buttons in the browser now work in an intuitive and
consistent way.  Additionally, this allows deep-linking specific image
within gallery.

The web server must be set up in the 'front controller' fashion for this
to work, though.  Following is the configuration that achieve required
front controller set up:

  RewriteEngine On
  RewriteBase "/path/to/galleries/"
  RewriteRule "\.json$" - [L]
  RewriteCond %{REQUEST_FILENAME} !-f
  RewriteRule . /path/to/gallery.html [L]

This is applied to <Directory> section '/path/to/galleries'; every
directory in that location will be examined for presence of the
index.json file.  The /path/to/gallery.html points to the gallery
skeleton file (part of this project).  Note that requests for JSON files
must be passed through unnchanged, so that the client-side can properly
detect it's presence or absence.